Samurai sudoku solver, and starter code
Sudoku.py is the starter code I used to write the samurai solver. It solves a regular sudoku. Running samurai.py gives you a prompt to enter a file containing a plain text sudoku in the format Bryan has published. Right now backtracking search is implemented, but forward checking and GAC are not.